msgr: bug fixes and improvements
* msgr: fix inbound connection bug (I wasn't setting conn->sock, oops)

* msg/fast_log.c: nicer fast_log printout

* msgr: add MCONN_AWAITING_HDR state

* msgr_unit: don't yak on stderr unless there is an error

+++ msg/msgr.h
@@ -89,7 +89,7 @@ extern struct msgr *msgr_init(char *err, size_t err_len,
  *
  * @return		the messenger on success; NULL otherwise
  */
-extern void msgr_listen(struct msgr *msgr, int port,
+extern void msgr_listen(struct msgr *msgr, uint16_t port,
 		char *err, size_t err_len);
 
 /** Start the messenger.
